+++ src/clack/pytest_plugin.py
@@ -16,13 +16,14 @@ from typing import Any, Protocol, Type
 from _pytest.fixtures import SubRequest
 from pytest import fixture
 
-from ._config_file import ConfigFile, YAMLConfigFile
+from ._config_file import YAMLConfigFile
+from .types import ClackConfigFile
 
 
 class MakeConfigFile(Protocol):
     """Type of the function returned by `make_config_file()`."""
 
-    def __call__(self, basename: str, **kwargs: Any) -> ConfigFile:
+    def __call__(self, basename: str, **kwargs: Any) -> ClackConfigFile:
         """Captures the `make_config_file()` function's signature."""
 
 
@@ -30,17 +31,17 @@ class MakeConfigFile(Protocol):
 def make_config_file_fixture(
     request: SubRequest, tmp_path: Path
 ) -> MakeConfigFile:
-    """Returns a function that can be used to generate ConfigFile objects.
+    """Returns a function that can be used to generate ClackConfigFile objects.
 
     The associated config file is first instantiated using the 'kwargs'
     provided by the caller.
     """
 
-    def make_config_file(basename: str, **kwargs: Any) -> ConfigFile:
+    def make_config_file(basename: str, **kwargs: Any) -> ClackConfigFile:
         path = tmp_path / basename
         path.parent.mkdir(parents=True, exist_ok=True)
 
-        config_file_type: Type[ConfigFile] = request.param
+        config_file_type: Type[ClackConfigFile] = request.param
         return config_file_type.new(path, **kwargs)
 
     return make_config_file