+++ tests/test_clack.py
@@ -12,7 +12,7 @@ from .shared import Config
 
 def test_new_command_factory() -> None:
     """Test the clack.new_command_factory() function."""
-    with dyn.clack_envvars_set("test_clack", [Config]):  # type: ignore[list-item]
+    with dyn.clack_envvars_set("test_clack", [Config]):
         parser = clack.Parser()
         new_command = clack.new_command_factory(parser, dest="command")
         foo = new_command("foo", help="Test FOO subcommand.")
@@ -27,7 +27,7 @@ def test_new_command_factory() -> None:
 
 def test_config_is_immutable() -> None:
     """Test that the Config object's attributes are immutable."""
-    with dyn.clack_envvars_set("test_clack", [Config]):  # type: ignore[list-item]
+    with dyn.clack_envvars_set("test_clack", [Config]):
         cfg = Config.from_cli_args(["", "--do-stuff"])
         assert cfg.do_stuff